Break down new project page into smaller components
