--- apps/studio/components/interfaces/Organization/NewProject/index.ts
@@ -1,1 +1,0 @@
-export { default as NotOrganizationOwnerWarning } from './NotOrganizationOwnerWarning'
