+++ apps/studio/components/interfaces/ProjectCreation/ProjectCreation.constants.ts
@@ -1,3 +1,5 @@
+import { DesiredInstanceSize } from 'data/projects/new-project.constants'
+
 // [Joshen] Obtained from https://gist.github.com/tadast/8827699
 export const COUNTRY_LAT_LON = {
   AF: { lat: 33, lon: 65 },
@@ -254,3 +256,5 @@ export const COUNTRY_LAT_LON = {
 }
 
 export const SPECIAL_CHARS_REGEX = /^[^@:\/]*$/
+
+export const sizes: DesiredInstanceSize[] = ['micro', 'small', 'medium']
