+++ apps/studio/data/organizations/keys.ts
@@ -1,5 +1,3 @@
-import type { DesiredInstanceSizeForAvailableRegions } from './organization-available-regions-query'
-
 export const organizationKeys = {
   list: () => ['organizations'] as const,
   detail: (slug?: string) => ['organizations', slug] as const,
@@ -22,9 +20,6 @@ export const organizationKeys = {
     ['organizations', slug, 'validate-token', token] as const,
   projectClaim: (slug: string, token: string) =>
     ['organizations', slug, 'project-claim', token] as const,
-  availableRegions: (
-    slug: string | undefined,
-    cloudProvider: string,
-    size?: DesiredInstanceSizeForAvailableRegions
-  ) => ['organizations', slug, 'available-regions', cloudProvider, size] as const,
+  availableRegions: (slug: string | undefined, cloudProvider: string, size?: string) =>
+    ['organizations', slug, 'available-regions', cloudProvider, size] as const,
 }
